requirements.txt
Flask==2.0.1
flask-cors==3.0.10
torch==1.11.0
torchvision==0.12.0
Pillow==9.0.1
gunicorn==20.1.0
requests==2.27.1